rebuild geom package.
some work left to do, but this is a good starting point.

big change is to make it possible to tag points and lines within a figure.  Also switch to immutable values by pointer for most things.  Use fewer interfaces.

big TODO is to implement transformations, but that's a task for another day.